Update control inventario de modulos

+++ src/main/java/com/irfeyal/servicio/inventarios/AprobacionKitService.java
@@ -33,6 +33,7 @@ import com.irfeyal.modelo.inventarios.TempPagoKit;
 import com.irfeyal.modelo.matricula.Estudiante;
 import com.irfeyal.modelo.rolseguridad.Empleado;
 import com.irfeyal.servicio.matricula.EstudianteServiceImpl;
+import com.irfeyal.servicio.rolseguridad.ModuloServices;
 
 @Service // ("IAutoServiceImplement")
 @Transactional
@@ -41,19 +42,62 @@ public class AprobacionKitService implements AprobacionKitInterface {
 	// @Qualifier("kitRepo")
 	@Autowired
 	private AprobacionKitDao aprobacionDao;
-	
+
 	@Autowired
 	private EstudianteServiceImpl estudianteServiceImpl;
 
 	@Autowired
 	private IKitService kitService;
 	
+	@Autowired
+	private ModulolibroService modulolibroService;
+
 	@Override
 	public AprobacionKit save(AprobacionKit aprobacion) {
 
-		AprobacionKit a = aprobacionDao.save(aprobacion);
+		List<Kit> listaKits = kitService.listAllKit();
+		List<ModuloLibro> listaModulos = modulolibroService.listAllModuloLibro();
+		List<ModuloLibro> listaRestarModulos = aprobacion.getKit().getListaModulos();
+
+		boolean validar = true;
+		for (int a = 0; a < listaKits.size(); a++) {
+			if (aprobacion.getKit().getId_kit() == listaKits.get(a).getId_kit()) {
+				for (int b = 0; b < listaModulos.size(); b++) {
+					for (int c = 0; c < listaRestarModulos.size(); c++) {
+						if (listaModulos.get(b).getId_modulo_libro() == listaRestarModulos.get(c).getId_modulo_libro()) {
+							if (listaModulos.get(b).getStockMinimo() > listaModulos.get(b).getCantidad()) {
+								validar = false;
+							} 
+						}
+					}
+				}
+				a = a + listaKits.size();
+			}
+		}
+
+		if(validar == true) {
+			for (int a = 0; a < listaKits.size(); a++) {
+				if (aprobacion.getKit().getId_kit() == listaKits.get(a).getId_kit()) {
+					for (int b = 0; b < listaModulos.size(); b++) {
+						for (int c = 0; c < listaRestarModulos.size(); c++) {
+							if (listaModulos.get(b).getId_modulo_libro() == listaRestarModulos.get(c).getId_modulo_libro()) {
+								ModuloLibro mD = modulolibroService.getById(listaRestarModulos.get(c).getId_modulo_libro()).get();
+								mD.setCantidad(mD.getCantidad() - 1);
+								modulolibroService.save(mD);
+							}
+						}
+					}
+					a = a + listaKits.size();
+				}
+			}
+			AprobacionKit a = aprobacionDao.save(aprobacion);
+
+			return a;
+		}else {
+			return null;
+		}
+
 		
-		return a;
 	}
 
 	@Override
@@ -74,73 +118,86 @@ public class AprobacionKitService implements AprobacionKitInterface {
 		return null;
 	}
 
-
-
-	
 	public boolean validarAprobacion(AprobacionKit aprobacion) {
 		String observacionAproba = aprobacion.getTipoAprobacion();
 		Boolean estadoAproba = aprobacion.getEstadoAproba();
 		String detalleControl = aprobacion.getDetalleControl();
 		Date fechaAprobacion = aprobacion.getFechaAprobacion();
-		
-		if(observacionAproba!=null 
-				&& estadoAproba != null
-				&& detalleControl != null
-				&& fechaAprobacion != null
-				&& estadoAproba != null
-				) {
+
+		if (observacionAproba != null && estadoAproba != null && detalleControl != null && fechaAprobacion != null
+				&& estadoAproba != null) {
 			return true;
 		}
-		
-		
+
 		return false;
 	}
 
 	@Override
 	public AprobacionKit delete(Long id_aprobacion) {
 		AprobacionKit aprobacion = getById(id_aprobacion).get();
-		if(aprobacion == null) {
+		if (aprobacion == null) {
 			return null;
-		}else {
-			aprobacionDao.deleteById(id_aprobacion);
-			return aprobacion;
-			
+		} else {
+			if(aprobacion.getEstadoAproba() == false) {
+				List<Kit> listaKits = kitService.listAllKit();
+				List<ModuloLibro> listaModulos = modulolibroService.listAllModuloLibro();
+				List<ModuloLibro> listaRestarModulos = aprobacion.getKit().getListaModulos();
+					for (int a = 0; a < listaKits.size(); a++) {
+						if (aprobacion.getKit().getId_kit() == listaKits.get(a).getId_kit()) {
+							for (int b = 0; b < listaModulos.size(); b++) {
+								for (int c = 0; c < listaRestarModulos.size(); c++) {
+									if (listaModulos.get(b).getId_modulo_libro() == listaRestarModulos.get(c).getId_modulo_libro()) {
+										ModuloLibro mD = modulolibroService.getById(listaRestarModulos.get(c).getId_modulo_libro()).get();
+										mD.setCantidad(mD.getCantidad() + 1);
+										modulolibroService.save(mD);
+									}
+								}
+							}
+							a = a + listaKits.size();
+						}
+						
+					}
+				aprobacionDao.deleteById(id_aprobacion);
+				return aprobacion;
+			}else {
+				return null;
+			}
+
 		}
 	}
-	
-	
+
 	public List<EstudiantePagoKit> listaPagosEstud() {
 		Gson gson = new Gson();
 		List<Object> obj = aprobacionDao.estudiantesPagado();
 		List<EstudiantePagoKit> listaEstudPagos = new ArrayList<>();
-		
+
 		for (Iterator iterator = obj.iterator(); iterator.hasNext();) {
 			Object object = (Object) iterator.next();
 			String a = gson.toJson(object);
 			JsonParser parseJSON = new JsonParser();
 			JsonArray stringlistatiposPago = (JsonArray) parseJSON.parse(a);
 			EstudiantePagoKit tempPagoKit = new EstudiantePagoKit();
-			tempPagoKit.setIdEstudiante(Integer.parseInt(""+stringlistatiposPago.get(0)));
+			tempPagoKit.setIdEstudiante(Integer.parseInt("" + stringlistatiposPago.get(0)));
 			tempPagoKit.setConceptoPago("Kit");
-			tempPagoKit.setIdKit(Integer.parseInt(""+stringlistatiposPago.get(2)));
-			tempPagoKit.setValorPagado(Float.parseFloat(""+stringlistatiposPago.get(3)));
+			tempPagoKit.setIdKit(Integer.parseInt("" + stringlistatiposPago.get(2)));
+			tempPagoKit.setValorPagado(Float.parseFloat("" + stringlistatiposPago.get(3)));
 			List<Estudiante> listEstudiante = estudianteServiceImpl.findAll();
 			for (int i = 0; i < listEstudiante.size(); i++) {
-				if((""+listEstudiante.get(i).getid_estudiante()).equals(""+tempPagoKit.getIdEstudiante())) {
+				if (("" + listEstudiante.get(i).getid_estudiante()).equals("" + tempPagoKit.getIdEstudiante())) {
 					tempPagoKit.setEstudiante(listEstudiante.get(i));
-					i=listEstudiante.size();
+					i = listEstudiante.size();
 				}
 			}
 			List<Kit> listaKit = kitService.listAllKit();
 			for (int i = 0; i < listaKit.size(); i++) {
-				if((""+listaKit.get(i).getId_kit()).equals(tempPagoKit.getIdKit()+"")) {
+				if (("" + listaKit.get(i).getId_kit()).equals(tempPagoKit.getIdKit() + "")) {
 					tempPagoKit.setKit(listaKit.get(i));
 				}
 			}
 			listaEstudPagos.add(tempPagoKit);
 		}
 		return listaEstudPagos;
-		
+
 	}
 
 }
